Remove the 'check' option
Since this option is only a very minimal hack to avoid actually parsing
YAML files and we can actually do this now there is no need to continue
to support this function in the future. It only supports a very basic
usecase which is completely covered by the way more flexible `verify`
option anyway.

No need to confuse users with options we don't want them to be using
anyway.